+++ server/src/main/java/org/allaymc/server/entity/component/player/EntityPlayerPhysicsComponentImpl.java
@@ -3,7 +3,7 @@ package org.allaymc.server.entity.component.player;
 import org.allaymc.api.entity.component.player.EntityPlayerNetworkComponent;
 import org.allaymc.api.eventbus.EventHandler;
 import org.allaymc.server.component.annotation.Dependency;
-import org.allaymc.server.entity.component.EntityPhysicsComponentImpl;
+import org.allaymc.server.entity.component.EntityHumanPhysicsComponentImpl;
 import org.allaymc.server.entity.component.event.CPlayerGameTypeChangeEvent;
 import org.cloudburstmc.math.vector.Vector3f;
 import org.cloudburstmc.protocol.bedrock.data.GameType;
@@ -14,7 +14,7 @@ import org.joml.Vector3dc;
 /**
  * @author daoge_cmd
  */
-public class EntityPlayerPhysicsComponentImpl extends EntityPhysicsComponentImpl {
+public class EntityPlayerPhysicsComponentImpl extends EntityHumanPhysicsComponentImpl {
 
     @Dependency
     protected EntityPlayerNetworkComponent networkComponent;